src/requirements.txt
asgiref==3.5.0
Django==3.2.12
django-crispy-forms==1.14.0
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.1.1
